feat(nuxt): expose `hashKey` utility for cache key hashing
