+++ packages/nuxt/src/app/composables/fetch.ts
@@ -3,7 +3,7 @@ import type { $Fetch, NitroFetchRequest, TypedInternalResponse, AvailableRouterM
 import type { MaybeRef, MaybeRefOrGetter, Ref } from 'vue'
 import { computed, reactive, toValue, watch } from 'vue'
 import { isPlainObject } from '@vue/shared'
-import { hashKey } from './hash'
+import { hashKey } from '../utils/hash'
 import type { AsyncData, AsyncDataOptions, KeysOf, MultiWatchSources, PickFrom, _Transform } from './asyncData'
 import { useAsyncData } from './asyncData'
 import type { NuxtError } from './error'
